requirements.txt
PyYAML==6.0.1
pandas==2.0.0
gradio==3.42.0
openai==0.28.0
psycopg2==2.9.7
langchain==0.0.284
gradio_client==0.5.0
langchain-experimental==0.0.15